updated version

+++ server/db/models/product.js
@@ -17,7 +17,7 @@ const Product = db.define('product', {
     }
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
     allowNull: false,
     validate: {
       notEmpty: true
@@ -40,4 +40,5 @@ const Product = db.define('product', {
   }
 })
 
+//hook
 module.exports = Product
